refactor: move signature extraction into code_audit to break refactor cycle
Grammar-driven method-signature extraction (`MethodSignature`,
`extract_signatures`, `extract_signatures_from_items`, `extract_method_body`)
lived in `refactor::plan::generate::signatures`, but `code_audit`'s convention
detector called up into it — while the extraction itself calls *down* into
`code_audit::core_fingerprint`. That formed a code_audit<->refactor production
cycle around the same neighborhood.

Extraction is fundamentally a source-parsing primitive over the audit
fingerprinting grammar, so it belongs in code_audit. Move it into a new
`code_audit::signature_extraction` module (next to `core_fingerprint`, which it
already depends on). The convention detector now calls the local module;
refactor's convention-fix codegen imports `MethodSignature` /
`extract_signatures_from_items` *down* from code_audit. The codegen helpers
(`generate_method_stub`, `generate_fallback_signature`, `parse_items_for_dedup`,
etc.) stay in refactor.

Removes the `code_audit::conventions -> refactor` edge. The remaining
code_audit->refactor edges (audit fixability planning in report.rs) are a
separate, larger cluster.

Refs #8425
